@@ -11,9 +11,9 @@
   "freshr": {},
   "dependencies": {
     "graphlib": "^2.1.7",
-    "rollup": "^1.12.4",
+    "rollup": "^2.21.0",
     "rxjs": "^6.5.2",
-    "sass": "^1.26.8",
-    "typescript": "3.8"
+    "sass": "^1.26.10",
+    "typescript": "3.9.6"
   }
 }
