@@ -25,8 +25,8 @@
   "dependencies": {
     "graphlib": "^2.1.7",
     "node-sass": "^4.12.0",
-    "rollup": "^1.12.4",
+    "rollup": "^2.21.0",
     "rxjs": "^6.5.2",
-    "typescript": "3.8"
+    "typescript": "3.9.6"
   }
 }
