@@ -14,7 +14,7 @@
     ]
   },
   "dependencies": {
-    "typescript": "3.8"
+    "typescript": "3.9.6"
   },
   "devDependencies": {
     "@types/react-dom": "^16.9.4"
