@@ -1,21 +1,29 @@
 import * as Path from 'path';
+
 import * as Express from 'express';
 import * as Websocket from 'socket.io';
-import { concat, EMPTY } from 'rxjs';
-import { map, mergeMap, tap } from 'rxjs/operators';
+import {concat, EMPTY} from 'rxjs';
+import {map, mergeMap, tap} from 'rxjs/operators';
 
-import { Log } from 'freshlog';
+import {Log} from 'freshlog';
+import {File, Freshr, Hypermedia, loadFiles, server} from 'freshr';
 
-import { server, Hypermedia, Processor, Plugin, Freshr, BuildManager, BuildStep, TaskDefinition, loadFiles, File } from 'freshr';
-
-const rollupAlias = require('rollup-plugin-alias');
-const rollupCommonjs = require('rollup-plugin-commonjs');
-const rollupResolve = require('rollup-plugin-node-resolve');
+import {buildSteps} from './build';
 
 Log.handlers.get('trace')!.enabled = true;
 
 const app = Express();
 const websocketServer = Websocket();
+
+// interface Config {
+    // sitePath: string;
+// }
+
+// const defaultConfig = {
+// };
+
+// const config = {
+// };
 
 
 const demoPath = Path.join(__dirname, '..', '..', 'client');
@@ -154,114 +162,6 @@ concat(
         const watcher = freshr.watchResources(sitePath);
         watcher.events.subscribe();
         // freshr.watcher.add(sitePath);
-
-        const buildSteps: BuildStep = {
-            sType: 'multitask',
-            sync: true,
-            steps: [{
-                sType: 'task',
-                definition: 'clean',
-                files: [{
-                    inputs: {target: ['build']},
-                    outputs: {}
-                }]
-            }, {
-                sType: 'multitask',
-                steps: [{
-                    sType: 'task',
-                    definition: 'copy',
-                    files: [{
-                        inputs: {target: [Path.join('src', 'partials')]},
-                        outputs: {destination: [Path.join('build', 'partials')]}
-                    }, {
-                        inputs: {target: [Path.join('src', 'site')]},
-                        outputs: {destination: [Path.join('build', 'site')]}
-                    }, {
-                        inputs: {target: [Path.join('src', 'templates')]},
-                        outputs: {destination: [Path.join('build', 'templates')]}
-                    }, {
-                        inputs: {target: [Path.join('..', '..', 'plugins', 'dashboard', 'build', 'components')]},
-                        outputs: {destination: [Path.join('build', 'js', '~dashboard')]}
-                    }, {
-                        inputs: {target: [Path.join('..', '..', 'plugins', 'material-design-icons', 'build')]},
-                        outputs: {destination: [Path.join('build', 'material-icons')]}
-                    }]
-                },  {
-                    sType: 'task',
-                    definition: 'sass',
-                    options: {
-                        includePaths: [Path.join(__dirname, '..', 'node_modules')]
-                    },
-                    files: [{
-                        inputs: {target: ['src/sass/freshr.scss']},
-                        outputs: {
-                            css: ['build/css/freshr.css'],
-                            sourceMap: ['build/css/freshr.css.map'],
-                        }
-                    }]
-                }, {
-                    sType: 'task',
-                    definition: 'sass',
-                    options: {
-                        includePaths: [Path.join(__dirname, '..', 'node_modules')]
-                    },
-                    files: [{
-                        inputs: {target: ['src/sass/dashboard.scss']},
-                        outputs: {
-                            css: ['build/css/dashboard.css'],
-                            sourceMap: ['build/css/dashboard.css.map'],
-                        }
-                    }]
-                }, {
-                    sType: 'task',
-                    definition: 'rollup',
-                    options: {
-                        rollup: {
-                            plugins: [
-                                rollupCommonjs({
-                                    include: [
-                                        'node_modules/**',
-                                    ],
-                                }),
-                                rollupResolve({
-                                    browser: true,
-                                    extensions: ['.js', '.jsx'],
-                                }),
-
-                                // rollupAlias({
-                                //     'cytoscape': Path.join(__dirname, '..', 'node_modules/cytoscape/dist/cytoscape.umd.js'),
-                                //     'dagre': Path.join(__dirname, '..', 'node_modules/dagre/dist/dagre.min.js'),
-                                //     'cytoscape-dagre': Path.join(__dirname, '..', 'node_modules/cytoscape-dagre/cytoscape-dagre.js'),
-                                //     'cytoscape-dagre': Path.join(__dirname, '..', 'node_modules/cytoscape-dagre/cytoscape-dagre.js'),
-                                // })
-                            ]
-                        }
-                    },
-                    files: [{
-                        inputs: {target: ['src/jsx/resource-graph.jsx']},
-                        outputs: {
-                            js: ['build/js/resource-graph.js'],
-                        }
-                    }]
-                }, {
-                    sType: 'task',
-                    definition: 'react-rollup',
-                    options: {
-                    },
-                    files: [{
-                        inputs: {target: ['src/jsx/main.jsx']},
-                        outputs: {
-                            js: ['build/js/main.js'],
-                        }
-                    }, {
-                        inputs: {target: ['src/jsx/~config/main.jsx']},
-                        outputs: {
-                            js: ['build/js/main.js'],
-                        }
-                    }]
-                }]
-            }]
-        };
 
         freshr.build.buildSteps['demo'] = buildSteps;
 
