@@ -1,144 +1,10 @@
-import { json as graphJson } from 'graphlib';
-import { Plugin, HAL, createSchema, objectDifference, Hypermedia, Embed, filterCuries, getProfiles, resourceMatchesProfile, Processor, ProcessorFn } from 'freshr';
+import { ProcessorFactory } from 'freshr';
 
+import { embed } from './embed';
 import { makeIndex } from './make-index';
 import { tags } from './tags';
 
-type ResourceState = Hypermedia.ResourceState;
-type ExtendedResource = Hypermedia.ExtendedResource;
-type CalculateFromResourceParams = Hypermedia.CalculateFromResourceParams;
-
-const processors: {[name: string]: Plugin.ProcessorGenerator} = {
-    self: () => ({
-        name: 'self',
-       fn: (rs: Hypermedia.ResourceState): Hypermedia.ResourceState => (
-            Object.assign(rs, {
-                resource: Object.assign(rs.resource, {
-                    _links: Object.assign({
-                        // self: {href: rs.state.baseUri? Url.resolve(rs.state.baseUri, rs.relativeUri): rs.relativeUri}
-                        self: {href: rs.relativeUri}
-                    }, rs.resource._links)
-                })
-            })
-        )
-    }),
-    /** try to auto-generate missing schema
-     *
-     * If a schema is missing and can be generated, the schema resource is initialized with a blueprint it will use to generate the full schema when it is processed
-     *
-     * schema generation steps:
-     *  1. if the resource only has one profile: create schema describing all fields and links in the resource, excluding self and profile links. 
-         u  2. if the resource has multiple profiles: if all profiles are already defined except one, create a schema describing fields and links that are not defined in the other profile schemas (set difference).
-     *  3. if the resource has multiple undefined profiles: emit a warning and do not generate any schemas.
-     * */
-    schema: () => ({
-        name: 'schema',
-        fn: (rs) => {
-            const ignoredSchemas = ['/schema/freshr/resource-graph'];
-            const schemaProfile = '/schema';
-            if(rs.relativeUri.startsWith(schemaProfile)) {
-                // generate the schema if it was initialized with a schema-source
-                const schemaSourceLink = rs.resource._links && rs.resource._links['fs:schema-source'] as HAL.Link | undefined;
-                if(schemaSourceLink) {
-                    const schemaSource = rs.calculateFrom(schemaSourceLink.href, ({resource}) => resource);
-                    const otherSchemaHrefs = getProfiles(schemaSource)
-                        .map((profile) => profile.href)
-                        .filter((href) => rs.hypermedia.normalizeUri(href) !== rs.relativeUri);
-
-                    const otherSchemas = rs.calculateFrom(otherSchemaHrefs, (r) => r.filter(({resource}) => resource));
-                    const prunedObject = otherSchemas.reduce((obj: ExtendedResource, schema: CalculateFromResourceParams) => {
-                        return objectDifference(obj, schema.resource!.schema);
-                    }, schemaSource);
-
-                    const schema = createSchema(prunedObject);
-                    return {
-                        ...rs,
-                        resource: {
-                            ...rs.resource,
-                            schema,
-                        }
-                    };
-                }
-            }
-            else {
-                // initialize a schema if there isn't already one and we can generate it, or emit a warning if the resource does not conform to the schema
-                // TODO: add option to disable schema validation
-                const profileHrefs = getProfiles(rs.resource).map((p) => p.href);
-                const missingProfiles = profileHrefs;
-                // const missingProfiles = rs.calculateFrom(profileHrefs, (r) => r.filter(({resource}) => !resource).map(({href}) => href));
-                if(missingProfiles.length === 0) {
-                    // TODO: validate the schema
-                }
-                else if(missingProfiles.length === 1) {
-                    // only one missing schema, initialize so it will be auto-generated
-                    if(!ignoredSchemas.includes(missingProfiles[0])) {
-                        rs.markDirty(missingProfiles[0], {
-                            "_links": {
-                                "profile": {
-                                    "href": schemaProfile
-                                },
-                                "fs:schema-source": {
-                                    "href": rs.relativeUri
-                                }
-                            }
-                        });
-                    }
-                }
-                else {
-                    // TODO: add logging capabilities to processors
-                    console.log(JSON.stringify({
-                        level: 'warn',
-                        message: `processor.schema: ${rs.relativeUri} has ${missingProfiles.length} unspecified profiles: ${missingProfiles.join(', ')}`
-                    }));
-                }
-
-            }
-            return rs;
-        }
-    }),
-
-    /** updates the resource graph resource */
-    resourceGraph: () => ({
-        name: 'resourceGraph',
-        fn: (rs) => {
-            const resourceGraphProfile = '/schema/freshr/resource-graph';
-            const resourceGraphUri = '/freshr/resource-graph';
-            if(resourceMatchesProfile(rs.resource, resourceGraphProfile)) {
-                const graph = graphJson.write(rs.state.resourceGraph) as any;
-                // create a copy of the resource-graph node without the "resource" property, to prevent circular reference
-                graph.nodes = graph.nodes.map((node: any) => {
-                    if(node.v !== `${resourceGraphUri}.json`) { // TODO: uri normalization
-                        return node;
-                    }
-
-                    const newNodeValue = Object.assign({}, node.value);
-                    delete newNodeValue.resource;
-
-                    return {
-                        ...node,
-                        value: newNodeValue
-                    };
-                });
-                return {
-                    ...rs,
-                    resource: {
-                        ...rs.resource,
-                        graph
-                    }
-                };
-            }
-
-            rs.markDirty(resourceGraphUri, {
-                "_links": {
-                    "profile": {
-                        "href": resourceGraphProfile
-                    }
-                }
-            });
-            return rs;
-        }
-    }),
-
+const processorFactories: {[name: string]: ProcessorFactory} = {
     // TODO: detect rels that use curies that haven't been defined
     // TODO: record local curie rels so we can generate warnings for rels that have no documentation resource */
     curies: () => ({
@@ -155,6 +21,7 @@ const processors: {[name: string]: Plugin.ProcessorGenerator} = {
             }
     }),
 
+    /*
     breadcrumb: () => ({
         name: 'breadcrumb',
         fn: (rs: ResourceState): ResourceState => {
@@ -172,6 +39,7 @@ const processors: {[name: string]: Plugin.ProcessorGenerator} = {
             return rs;
         }
     }),
+    */
 
     /**
      * add resources to the "_embedded" property for each rel in the "_embed" property. Then remove "_embed"
@@ -182,86 +50,11 @@ const processors: {[name: string]: Plugin.ProcessorGenerator} = {
      */
     embed: () => ({
         name: 'embed',
-        fn: (rs: ResourceState): ResourceState => {
-            const _embed: Embed = rs.resource._embed;
-            if(!_embed) {
-                return rs;
-            }
-
-            let resource = Object.assign({}, rs.resource);
-
-            const embedded = Object.keys(_embed).reduce((embedded, embedRel) => {
-                const embedEntry = _embed[embedRel];
-
-                // collect hrefs to embed
-                const embedHrefs = embedEntry.href && (Array.isArray(embedEntry.href)? embedEntry.href: [embedEntry.href]) || [];
-
-                const links = rs.resource._links && rs.resource._links[embedRel];
-                const linkHrefs = links && (Array.isArray(links)? links: [links]).map((link) => link.href) || [];
-
-                let hrefs = embedHrefs.concat(linkHrefs);
-                let linkHrefsUsed = linkHrefs.length;
-                if(embedEntry.max) {
-                    // embed hrefs are always used first, and linkHrefs are used in order, so we can easily calculate how many link hrefs were used (and should be deleted)
-                    linkHrefsUsed = Math.max(0, linkHrefs.length - Math.max(0, hrefs.length - embedEntry.max));
-                    hrefs = hrefs.slice(0, embedEntry.max);
-                }
-
-                embedded[embedRel] = rs.calculateFrom(hrefs, (resourceParams) => {
-                    return resourceParams.map(({resource: r, href}) => {
-                        // TODO: do something reasonable if the document to embed does not exist
-                        if(!r) {
-                            return r;
-                        }
-
-                        // TODO: implement depth
-                        if(embedEntry.properties) {
-                            return Object.assign(
-                                // always include the "self" link
-                                {_links: {self: r._links && r._links.self || href}},
-                                embedEntry.properties.reduce((obj, property) => {
-                                    if(property in r) {
-                                        obj[property] = r[property];
-                                    }
-                                    return obj;
-                                }, {} as ExtendedResource));
-                        }
-                        return r;
-                    });
-                }).filter((r?: HAL.Resource) => !!r);
-
-                // special case: if we got the href from a non-array "_link" entry, and there were no additional hrefs in "_embed", don't make the embedded resource an array.
-                // this simplifies template code when you were only expecting a single link
-                if(linkHrefs.length === 1 && embedded[embedRel].length === 1) {
-                    embedded[embedRel] = embedded[embedRel][0];
-                }
-
-                // delete links that were embedded
-                // TODO: don't delete links to resources that couldnt' be loaded?
-                if(linkHrefsUsed > 0) {
-                    if(!Array.isArray(links) || linkHrefsUsed === links.length) {
-                        delete resource._links![embedRel];
-                    }
-                    else {
-                        resource._links![embedRel] = links.slice(linkHrefsUsed);
-                    }
-                }
-
-                return embedded;
-            }, {} as any);
-
-            delete resource._embed;
-
-            return Object.assign(rs, {
-                resource: Object.assign(resource, {_embedded: embedded})
-            });
-        }
+        fn: embed,
     }),
     makeIndex,
     tags: () => tags,
-}
-
-export default processors;
+};
 
 /* higher-order processor that only runs the provided processor if the resource matches the designated profile */
 export const matchProfile = (profile: HAL.Uri, processor: Processor): Processor => {
@@ -272,3 +65,5 @@ export const matchProfile = (profile: HAL.Uri, processor: Processor): Processor 
             rs
     };
 };
+
+export { processorFactories };
